requirements.txt
asgiref==3.2.10
Django==3.0.8
django-cors-headers==3.4.0
django-extensions==3.0.3
django-rest-framework==0.1.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
PyJWT==1.7.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
APScheduler~=3.6.3
djangorestframework-jwt~=1.11.0